Initial Commit
Created script to target Kyocera scanners to steal SMB credentials.  Admin password for device is needed (to access address book).  Still need to do: scan entire address book for entries saving to SMB.  As of now, assuming that the first entry is a folder.  Probably will cause failure if this is not the case.